refactor(mobile): simplify notification code
